+++ cli/main.py
@@ -10,6 +10,7 @@ import click
 from cli.account_balance_cmd import account_balance
 from cli.close_books_cmd import close_books
 from cli.delete_transaction_cmd import delete_transaction_by_guid
+from cli.export_accounts_cmd import export_accounts
 from cli.export_beancount_cmd import export_beancount
 from cli.export_cmd import export_transactions
 from cli.export_transaction_cmd import export_transaction
@@ -54,6 +55,7 @@ def run(script, args):
 
 # Register commands
 cli.add_command(export_transactions, name='export')
+cli.add_command(export_accounts, name='export-accounts')
 cli.add_command(import_transactions, name='import')
 cli.add_command(validate_ledger, name='validate')
 cli.add_command(export_beancount, name='export-beancount')